TECH-4_2: Fixes to clinic admin profile.

+++ client/src/app/components/adminC-dashboard/doctor-listing/doctor-listing.component.ts
@@ -3,6 +3,7 @@ import { ClinicService } from 'src/app/services/clinic.service';
 import { Router, ActivatedRoute, NavigationEnd } from '@angular/router';
 import { UserService } from 'src/app/services/user.service';
 import { FormGroup, FormBuilder } from '@angular/forms';
+import { DialogService } from 'src/app/services/dialog.service';
 
 @Component({
 	selector: 'app-doctor-listing',
@@ -23,7 +24,8 @@ export class DoctorListingComponent implements OnInit {
 		private router: Router,
 		private activatedRoute: ActivatedRoute,
 		private userService: UserService,
-		private formBuilder: FormBuilder) {
+		private formBuilder: FormBuilder,
+		private confirmationDialogService: DialogService) {
 		this.navigationSubscription = this.router.events.subscribe((e: any) => {
 			if (e instanceof NavigationEnd) {
 				this.clinicID = this.userService.getUser().clinic_id;
@@ -61,14 +63,19 @@ export class DoctorListingComponent implements OnInit {
 	}
 
 	deleteDoctor(doctor) {
-		this.clinicService.deleteDoctor(doctor.id).subscribe(
-			(data) => {
-				this.getDoctors();
-			},
-			(error) => {
-				alert(error);
+		this.confirmationDialogService.confirm('Please confirm', 'Are you sure you want to delete doctor: ' + doctor.name+' '+doctor.surname + ' ?', false)
+		.then((confirmed) => {
+			if (confirmed.submited) {
+				this.clinicService.deleteDoctor(doctor.id).subscribe(
+					(data) => {
+						this.getDoctors();
+					},
+					(error) => {
+						alert(error);
+					}
+				)
 			}
-		)
+		});
 	}
 
 	onSearch(){
